+++ Sources/HabitatCore/Models.swift
@@ -420,6 +420,7 @@ public enum ValidationCommandPurpose: String, Codable, Equatable {
     case ordinaryLocal = "ordinary_local"
     case releaseArtifact = "release_artifact"
     case deviceVerification = "device_verification"
+    case environmentCheck = "environment_check"
 }
 
 public struct ValidationCommandClaim: Codable, Equatable {